addded a fourth alternative for internet
After finding the Tethered phone approach to be inconsistent (sometimes `ifuse` would see the phone, sometimes not), and the built-in wireless failed to work, I went out today to buy the "Apple proprietary ethernet thing".  But you cannot seem to get it from the Apple stores, at least not the physical ones; they've moved on to Thunderbolt-oriented devices alone, since that's what the current Air models supply.

I looked for a USB <-> ethernet adapter at another store, but could not find one.

What I *did* find was a third-party (D-link) USB to internet adapter.  I quickly googled the model numbers, figured out which one had built-in support via the site linked above, and bought it for about $12.

And huzzah, this has led to a much smoother installation experience for me.

So I figured I would suggest adding it to the list above.